+++ build.js
@@ -31,9 +31,16 @@ const prompt = require('prompt');
 const fs = require('fs');
 const request = require('request');
 const async = require('async');
+const path = require('path');
 const template = fs.readFileSync(__dirname + '/phantom/template.html').toString();
 const package = require(__dirname + '/package.json');
-const cdnURL = 'https://code.highcharts.com/'
+
+let npmLocation = __dirname + '/node_modules/highcharts/';
+
+let highchartsPackage = {};
+let officialCDNURL = 'https://code.highcharts.com/';
+let cdnURL = officialCDNURL;
+let useNPM = true;
 
 // We allow the fetch for these to fail without error.
 // This is because it's only available in version 6+
@@ -41,8 +48,32 @@ const cdnScriptsOptional = {
   '{{version}}/modules/sunburst.js': 1,
   '{{version}}/modules/xrange.js': 1,
   '{{version}}/modules/streamgraph.js': 1,
-  '{{version}}/modules/tilemap.js': 1
+  '{{version}}/modules/sankey.js': 1,
+  '{{version}}/modules/tilemap.js': 1,
+  '{{version}}/modules/histogram-bellcurve.js': 1,
+  '{{version}}/modules/bullet.js': 1,
+  '{{version}}/modules/organization.js': 1,
+  '{{version}}/modules/funnel3d.js': 1,
+  '{{version}}/modules/pyramid3d.js': 1,
+  '{{version}}/modules/dependency-wheel.js': 1,
+  '{{version}}/modules/item-series.js': 1,
+  '{{version}}/modules/timeline.js': 1,
+  '{{version}}/modules/pareto.js': 1,
+  '{{version}}/modules/coloraxis.js': 1,
+  '{{version}}/modules/venn.js': 1
 };
+
+// The scripts here will appear as user prompts
+const cdnScriptsQuery = {
+  "wordcloud": "{{version}}/modules/wordcloud.js",
+  "annotations": "{{version}}/modules/annotations.js"
+};
+
+// Push raw URL's here to force include them
+const cdnAdditional = [];
+
+// Push map collection includes here
+const cdnMapCollection = [];
 
 const cdnScriptsCommon = [
     "{{version}}/highcharts-3d.js",
@@ -72,10 +103,32 @@ const cdnLegacy = [
 ];
 
 const cdnMaps = [
-    "maps/{{version}}/modules/map.js"
+    'maps/{{version}}/modules/map.js'
 ];
 
+const cdnGantt = [
+    '{{version}}/modules/gantt.js'
+];
+
+const cdnMoment = [
+    'https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.18.1/moment.min.js',
+    'https://cdnjs.cloudflare.com/ajax/libs/moment-timezone/0.5.13/moment-timezone-with-data-2012-2022.min.js'
+];
+
+const rawScripts = [];
+
+const cachedScripts = {};
+
+
 ////////////////////////////////////////////////////////////////////////////////
+
+const boolConform = (value) => {
+  value = value.toUpperCase();
+  return value === 'Y'   ||
+         value === 'N'   ||
+         value === 'YES' ||
+         value === 'NO';
+};
 
 let schema = {
     properties: {
@@ -84,63 +137,72 @@ let schema = {
             required: true,
             default: 'no',
             message: 'Please enter (y)es or (n)o',
-            conform: function (value) {
-                value = value.toUpperCase();
-                return value === 'Y'   ||
-                       value === 'N'   ||
-                       value === 'YES' ||
-                       value === 'NO';
-            }
+            conform: boolConform
         },
         version: {
             description: 'Select your Highcharts version (e.g. 4.2.2):',
             required: true,
             message: 'Enter as e.g. 4.2.2. Default is latest.',
-            default: 'latest'
+            default: process.env.HIGHCHARTS_VERSION || 'latest'
         },
         maps: {
             description: 'Include Maps? (requires Maps license)',
             default: 'no',
             required: true,
-            conform: function (value) {
-                value = value.toUpperCase();
-                return value === 'Y'   ||
-                       value === 'N'   ||
-                       value === 'YES' ||
-                       value === 'NO'
-                ;
-            }
+            conform: boolConform
+        },
+        gantt: {
+          description: 'Include Gantt? (requires Gantt license, and >V6.2)',
+          default: 'no',
+          required: true,
+          conform: boolConform
         },
         styledMode: {
-            description: 'Enable styled mode? (requires Highcharts/Highstock 5 license)',
+            description: 'Enable styled mode? (requires Highcharts/Highstock 5+ license)',
             default: 'no',
             required: true,
-            conform: function (value) {
-                value = value.toUpperCase();
-                return value === 'Y'   ||
-                       value === 'N'   ||
-                       value === 'YES' ||
-                       value === 'NO'
-                ;
-            }
+            conform: boolConform
+        },
+        moment: {
+          description: 'Include moment.js for date/time handling?',
+          default: 'no',
+          required: true,
+          conform: boolConform
+        },
+        cdnURL: {
+          description: 'Which CDN would you like to use?',
+          default: process.env.HIGHCHARTS_CDN || cdnURL
         }
     }
 };
 
 ////////////////////////////////////////////////////////////////////////////////
 
+// Augment the schema with the query CDN scripts
+
+Object.keys(cdnScriptsQuery).forEach((name) => {
+  schema.properties[name] = {
+    description: `Enable ${name} support? y/n`,
+    required: false,
+    default: 'no',
+    conform: boolConform
+  };
+});
+
+////////////////////////////////////////////////////////////////////////////////
+
 require('colors');
 
-function embed(version, scripts, out, fn) {
+function embed(version, scripts, out, fn, optionals) {
     var funs = [],
         scriptBody = ''
     ;
 
+    optionals = optionals || {};
+
     if (version) {
         version = version.trim();
     }
-
-    console.log(version);
 
     if (version && parseInt(version[0]) < 5 && version[0] !== 'c')  {
         scripts = scripts.concat(cdnLegacy);
@@ -150,35 +212,44 @@ function embed(version, scripts, out, fn) {
         let scriptOriginal = script;
         let fullURL = '';
 
-        if (version !== 'latest' && version) {
+        if (version !== 'latest' && version && !useNPM) {
             script = script.replace('{{version}}', version);
         } else {
             script = script.replace('{{version}}/', '');
         }
 
         // Allow using full URLs in the include arrays
-        if (script.indexOf('http') >= 0) {
+        if (useNPM) {
+          fullURL = path.join(npmLocation, script.replace('stock/', ''));
+        } else if (script.indexOf('http') >= 0) {
           fullURL = script;
         } else {
           fullURL = cdnURL + script;
         }
 
-        console.log('  ', (fullURL).gray);
-
         funs.push(function (next) {
-            request(fullURL, function (error, response, body) {
+            // If we've allready fetched the required script, just return it.
+            if (cachedScripts[fullURL]) {
+              console.log(('   using cached fetch for ' + fullURL).gray);
+              scriptBody += cachedScripts[fullURL] + ';';
+              return next();
+            }
 
-                if (error) {
-                  if (cdnScriptsOptional[scriptOriginal]) {
-                    console.log(`  ${script} is not available for v${version}`.gray)
-                    return next();
-                  }
+            console.log('  ', (fullURL).gray);
 
-                  return next(error, fullURL);
-                }
+            const handleScript = (error, body) => {
 
-                if (body.trim().indexOf('<!DOCTYPE') === 0) {
-                  if (cdnScriptsOptional[scriptOriginal]) {
+                // if ((error || body.trim().indexOf('<!DOCTYPE') === 0)) {
+                //   console.log(optionals, scriptOriginal);
+                //   if (optionals[scriptOriginal]) {
+                //     console.log(`  notice: ${script} is not available for v${version}, skipped.`.yellow)
+                    // return next();
+                  // }
+                  // return next(error, fullURL);
+                // }
+
+                if ((body || '').trim().indexOf('<!DOCTYPE') === 0) {
+                  if (optionals[scriptOriginal]) {
                     console.log(`   ${script.substr(script.lastIndexOf('/') + 1)} is not available for v${version}, skipped..`.yellow);
                     return next();
                   }
@@ -186,9 +257,22 @@ function embed(version, scripts, out, fn) {
                   return next(404, script);
                 }
 
-                scriptBody += body;
+                cachedScripts[fullURL] = body;
+                scriptBody += body + ';\n';
                 next();
-            });
+            };
+
+            if (useNPM) {
+              // Fetch from npm package instead
+              script = script.replace('stock/', '');
+              fs.readFile(fullURL, 'utf8', (err, data) => {
+                handleScript(err, data);
+              });
+            } else {
+              request(fullURL, function (error, response, body) {
+                handleScript(error, body);
+              });
+            }
         });
     });
 
@@ -200,9 +284,13 @@ function embed(version, scripts, out, fn) {
             return startPrompt();
         }
 
-        if (err) {
-            return console.log('error fetching Highcharts:', err);
+        if (err && !useNPM) {
+            return console.log('error fetching Highcharts:', err, `
+            If you're behind a proxy, please follow this guide:
+            https://github.com/request/request#controlling-proxy-behaviour-using-environment-variables`);
         }
+
+        let additionalScripts = rawScripts.map((s) => `<script src="${s}"></script>`).join('') || '';
 
         console.log('Creating export template', out + '..');
 
@@ -211,6 +299,7 @@ function embed(version, scripts, out, fn) {
             template
                 .replace('"{{highcharts}}";', scriptBody)
                 .replace('<div style="padding:5px;">', '<div style="padding:5px;display:none;">')
+                .replace('{{additionalScripts}}', additionalScripts)
                 ,
             function (err) {
                 if (err) return console.log('Error creating template:', err);
@@ -225,58 +314,149 @@ function endMsg() {
     console.log('For documentation, see https://github.com/highcharts/node-export-server');
 }
 
-function embedAll(version, includeStyled, includeMaps) {
-    var standard = cdnScriptsStandard.concat(cdnScriptsCommon),
-        styled = cdnScriptsStyled.concat(cdnScriptsCommon)
+function embedAll(version, includeStyled, includeMaps, includeMoment, includeGantt, optionals) {
+    var standard = cdnScriptsStandard.concat(cdnScriptsCommon).concat(cdnAdditional),
+        styled = cdnScriptsStyled.concat(cdnScriptsCommon).concat(cdnAdditional)
     ;
 
+    optionals = optionals || {};
+
     if (includeMaps) {
-        standard = standard.concat(cdnMaps);
-        styled = standard.concat(cdnMaps);
+        console.log('Including maps support'.green);
+        standard = standard.concat(cdnMaps).concat(cdnMapCollection);
+        styled = styled.concat(cdnMaps).concat(cdnMapCollection);
+
+        // Map collections are user supplied, so we need to allow them to 404
+        cdnMapCollection.forEach((url) => {
+          optionals[url] = 1;
+        });
     }
 
-    console.log('Pulling Highcharts from CDN (' + version + ')..');
-    embed(version,
-          standard,
-          'export',
-          function () {
-            if (includeStyled) {
-                embed(false,
-                      styled,
-                      'export_styled',
-                      endMsg
-                );
-            } else {
-                endMsg();
-            }
+    if (includeGantt) {
+        console.log('Including Gantt support'.green);
+
+        standard = standard.concat(cdnGantt);
+        styled = styled.concat(cdnGantt);
+
+        // Gantt was introduced in 6.2. To avoid 404 errors if fetching an
+        // older version by accident, let the fetch fail gracefully
+        cdnGantt.forEach((url) => {
+            optionals[url] = 1
+        });
+    }
+
+    if (includeMoment) {
+        console.log('Including moment.js support'.green);
+        cdnMoment.forEach((t) => { rawScripts.push(t); });
+    }
+
+    useNPM = process.env.HIGHCHARTS_USE_NPM || cdnURL === 'npm';
+
+    if (useNPM) {
+      // Reset cdn url to official CDN in case we need to do a fallback
+      cdnURL = officialCDNURL;
+
+      // See if we need to check the parent folder instead
+      if (!fs.existsSync(npmLocation)) {
+        npmLocation = __dirname + '/../highcharts/';
+      }
+
+      // Check if the NPM package has been installed before doing anything.
+      if (!fs.existsSync(npmLocation)) {
+        console.log('Could not get Highcharts through NPM: the NPM package is not installed! Using fallback to CDN.'.red);
+        useNPM = false;
+      } else {
+        highchartsPackage = require(path.join(npmLocation, 'package.json'));
+        version = highchartsPackage.version || version;
+        console.log(`Bundling in Highcharts from NPM package @${highchartsPackage.version}..`.gray);
+      }
+    } else {
+      console.log(('Pulling Highcharts from CDN (' + version + ')..').gray);
+    }
+
+    embed(
+      version,
+      standard,
+      'export',
+      function () {
+        if (includeStyled) {
+            console.log('Including styled mode support'.green);
+            embed(
+              false,
+              styled,
+              'export_styled',
+              endMsg,
+              optionals
+            );
+        } else {
+         endMsg();
         }
+      },
+      optionals
     );
 }
 
+function affirmative(str) {
+  str = (str || '').toUpperCase();
+  return str === 'YES' || str === 'Y' || str === '1';
+}
+
+function getOptionals(include, forceInclude) {
+  let optionalScripts = {};
+
+  Object.keys(cdnScriptsOptional).forEach((url) => {
+    optionalScripts[url] = 1;
+  });
+
+  // Build list of optionals
+  Object.keys(cdnScriptsQuery).forEach((name) => {
+    if (forceInclude || (!include || affirmative(include[name]))) {
+      optionalScripts[cdnScriptsQuery[name]] = 1;
+      cdnAdditional.push(cdnScriptsQuery[name]);
+    }
+  });
+
+  return optionalScripts;
+}
+
 function startPrompt() {
-    prompt.message = '';
-    prompt.start();
+  prompt.message = '';
+  prompt.start();
 
-    prompt.get(schema, function (err, result) {
-        result.agree = result.agree.toUpperCase();
+  prompt.get(schema, function (err, result) {
+    result.agree = result.agree.toUpperCase();
 
-        if (result.agree === 'Y' || result.agree === 'YES') {
-            embedAll(result.version,
-                     result.styledMode.toUpperCase() === 'Y' ||
-                     result.styledMode.toUpperCase() === 'YES',
-                     result.maps.toUpperCase() === 'Y' ||
-                     result.maps.toUpperCase() === 'YES'
-            );
-        } else {
-            console.log('License terms not accepted, aborting'.red);
-        }
-    });
+    cdnURL = result.cdnURL || cdnURL;
+
+    if (result.agree === 'Y' || result.agree === 'YES') {
+        embedAll(result.version,
+                 affirmative(result.styledMode),
+                 affirmative(result.maps),
+                 affirmative(result.moment),
+                 affirmative(result.gantt),
+                 getOptionals(result)
+        );
+    } else {
+        console.log('License terms not accepted, aborting'.red);
+    }
+  });
+}
+
+function useIfDefined(what, def) {
+  return (typeof what === 'undefined' ? def : what);
 }
 
 if (process.env.ACCEPT_HIGHCHARTS_LICENSE) {
-    embedAll(process.env.HIGHCHARTS_VERSION || 'latest',
-             process.env.HIGHCHARTS_USE_STYLED || true,
-             process.env.HIGHCHARTS_USE_MAPS || true
+
+    cdnURL = process.env.HIGHCHARTS_CDN || cdnURL;
+
+    embedAll(
+      useIfDefined(process.env.HIGHCHARTS_VERSION, 'latest'),
+      useIfDefined(process.env.HIGHCHARTS_USE_STYLED, true),
+      useIfDefined(process.env.HIGHCHARTS_USE_MAPS, true),
+      useIfDefined(process.env.HIGHCHARTS_MOMENT, false),
+      useIfDefined(process.env.HIGHCHARTS_USE_GANTT, true),
+      getOptionals(cdnScriptsOptional, true)
     );
 } else {
     console.log(fs.readFileSync(__dirname + '/msg/licenseagree.msg').toString().bold);